requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apispec==6.8.1
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==45.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
Flask==3.1.0
Flask-Migrate==4.1.0
flask-smorest==0.46.0
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
webargs==8.7.0
Werkzeug==3.1.3
